Adding changes

+++ app/src/main/java/ca/georgebrown/comp3074/project/Event/Event.java
@@ -6,14 +6,14 @@ import ca.georgebrown.comp3074.project.Backpack.Backpack;
 import ca.georgebrown.comp3074.project.Route.Route;
 
 public class Event implements Serializable {
-    private int Event_Id;
+    private long Event_Id;
     private String Event_Name;
     private String Date;
     private String Description;
     private long backpack;
     private long route;
 
-    public Event(int id, String name, String date, String desc, long bp_id, long r_id)
+    public Event(long id, String name, String date, String desc, long bp_id, long r_id)
     {
         Event_Id = id;
         Event_Name = name;
@@ -22,7 +22,7 @@ public class Event implements Serializable {
         backpack = bp_id;
         route = r_id;
     }
-    public int getEvent_Id() {
+    public long getEvent_Id() {
         return Event_Id;
     }
